♻️ :: 리팩토링 반영
