+++ presentation/src/main/java/team/aliens/dms_android/feature/notice/NoticeDetailsScreen.kt
@@ -12,12 +12,12 @@ import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.Divider
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
 import java.util.UUID
 import team.aliens.design_system.theme.DormTheme
@@ -34,7 +34,7 @@ internal fun NoticeDetailsScreen(
     noticesViewModel: NoticesViewModel = hiltViewModel(),
 ) {
 
-    val state = noticesViewModel.uiState.collectAsState()
+    val state = noticesViewModel.uiState.collectAsStateWithLifecycle()
 
     val notice = state.value.notice
